Move site_packages helper test function to utils
